Seperate plugin options into sub-cmd
